osd: split out a backend-neutral core
Move Dear ImGui and the OSD logic (state machine, menu, file browser,
log view) into src/osd as a shared, frontend-neutral module. The SDL
frontend now drives this core through a thin glue layer instead of
owning the whole OSD itself.
