+++ src/osd/CMakeLists.txt
@@ -1,0 +1,27 @@
+#
+# 86Box    A hypervisor and IBM PC system emulator that specializes in
+#          running old operating systems and software designed for IBM
+#          PC systems and compatibles from 1981 through fairly recent
+#          system designs based on the PCI bus.
+#
+#          This file is part of the 86Box distribution.
+#
+#          CMake build script.
+#
+
+add_library(osd OBJECT
+    osd_core.cpp
+    imgui/imgui.cpp
+    imgui/imgui_draw.cpp
+    imgui/imgui_tables.cpp
+    imgui/imgui_widgets.cpp
+)
+
+set_target_properties(osd PROPERTIES AUTOMOC OFF AUTOUIC OFF AUTORCC OFF)
+
+target_include_directories(osd PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}
+    ${CMAKE_CURRENT_SOURCE_DIR}/imgui
+)
+
+target_link_libraries(86Box osd)
