+++ src/unix/sdl_shader.h
@@ -15,11 +15,7 @@ void sdl_shader_blit(SDL_Window *win, const void *pixels,
 void sdl_shader_close(void);
 int  sdl_shader_active(void);
 
-/* OSD compositing: render into fb_tex via FBO so shader applies to OSD */
-void          sdl_shader_begin_osd(int src_w, int src_h);
-void          sdl_shader_end_osd(void);
 SDL_GLContext sdl_shader_get_context(void);
-void          sdl_shader_get_viewport(int *dst_x, int *dst_y, int *dst_w, int *dst_h);
 
 #ifdef __cplusplus
 }
